\name{auteur}

\alias{auteur}
\alias{rjmcmc}

\docType{package}

\title{main: Bayesian Inference of Trait Evolutionary Process}

\description{
	\pkg{auteur} (\bold{A}ccommodating \bold{U}ncertainty in \bold{T}rait \bold{E}volution \bold{U}sing \bold{R}) performs reversible-jump Markov chain Monte Carlo sampling to identify phylogenetically localized
	shifts in the process of continuous trait evolution. 
}

\details{
 \tabular{ll}{
 Package: \tab auteur\cr
 Type: \tab Package\cr
 License: \tab GPL\cr
 LazyLoad: \tab yes\cr
 }
}

\author{
Luke J Harmon, Andrew L Hipp, Michael E Alfaro, Paul Joyce, and Jonathan M Eastman <jonathan.eastman@gmail.com>
}


