\name{geiger-internal}

\alias{print.bm}
\alias{print.mkn}
\alias{transform.phylo}

\title{internal geiger functions}
\description{\code{geiger-internal} functions are either not typically called by the user or are currently undocumented}

\details{
This is an internal \pkg{geiger} function, either not intended to be called directly by the user or is currently undocumented.
}
