\name{load.rjmcmc}
\alias{load.rjmcmc}
\title{processing posterior samples from single or multiple MCMC runs}
\description{Generates a pooled sample of posterior estimates from any number of independent MCMC runs}
\usage{

	load.rjmcmc(x, phy = NULL, burnin = NULL, thin = NULL)
}

\arguments{
  \item{x}{a vector of directory names (character strings) in which samples generated by \pkg{geiger} are found}
  \item{phy}{a phylogenetic tree of class 'phylo' for which to compile results; if NULL, the tree stored within each generated \code{rda} file is used}
  \item{burnin}{the proportion of each chain to be removed as burnin}
  \item{thin}{an integer specifying the thinning interval for chains}
}

\details{
This function provides a means to compile results from at least a single MCMC run.  It is intended to facilitate pooling of multiple independent runs. In cases where MCMC chains were sampled
using a set of (different) trees, the argument \code{phy} can be used to summarize compiled results against a single summary tree. Branchwise samples are stored with unique edge identifiers (character strings) 
that are more reliable than the use of numeric node labels (see \code{\link[digest]{digest}} for the function used to generate unique edge labels). 

If it is desired to run analyses for the same trait data across a set of trees (see \bold{Examples}), it is strongly recommended that the set of trees be sent to \code{rjmcmc.bm} as a \code{multiPhylo} object (see \code{\link[ape]{read.tree}}).
}  

\value{
an object of class \code{rjmcmc} or \code{rjmcmcmc}}

\author{JM Eastman}

\seealso{\code{\link[geiger]{rjmcmc.bm}}}

\examples{

a=ladderize(rcoal(30), right=FALSE)
dat=rTraitCont(a)
while(1){
	b=a
	b$tip.label[1:3]=sample(b$tip.label[1:3])
	if(!all(a$tip.label==b$tip.label)) break()
}

trees=list(a=a,b=b, c=a)
class(trees)="multiPhylo"
rjmcmc.bm(trees, dat, constrainJUMP=0)

res=load.rjmcmc(paste("relaxedBM", names(trees), sep="."), phy=a)
plot(res, par="shifts", show.tip=FALSE, edge.width=2.5)

}
