\name{geiger-internal}

\alias{print.bm}
\alias{print.mkn}
\alias{transform.phylo}
\alias{ROOT.ALL}
\alias{ROOT.BOTH}
\alias{ROOT.EQUI}
\alias{ROOT.FLAT}
\alias{ROOT.GIVEN}
\alias{ROOT.MAX}
\alias{ROOT.OBS}
\alias{acegram}
\alias{argn}
\alias{argn.default}
\alias{argn.bm}
\alias{argn.mkn}
\alias{argn<-}
\alias{argn<-.bm}
\alias{bf}
\alias{bm.lik}
\alias{constrain}
\alias{constrain.k}
\alias{constrain.m}
\alias{dlunif}
\alias{drop.tip}
\alias{dtpois}
\alias{edges.phylo}
\alias{edgelabels.auteur}
\alias{ess}
\alias{fitDiversification}
\alias{fitSplitModel}
\alias{gbcontain}
\alias{gbresolve.default}
\alias{gbresolve.phylo}
\alias{gbtaxdump}
\alias{gen}
\alias{get.simulation.matrix}
\alias{getBD}
\alias{getDiversificationModel}
\alias{getFullSplitModel}
\alias{glomogram.phylo}
\alias{hash.node}
\alias{hashes.phylo}
\alias{hashes.rjmcmc}
\alias{hdr}
\alias{heights}
\alias{heights.multiPhylo}
\alias{heights.phylo}
\alias{hme}
\alias{is.constrained}
\alias{is.phylo}
\alias{is.root}
\alias{kendallmoran.rate}
\alias{likfx.bm}
\alias{Linnaean}
\alias{load_taxdump}
\alias{make.bm.relaxed}
\alias{mkn.lik}
\alias{node.sons}
\alias{PATHd8.phylo}
\alias{picAnc}
\alias{plotDiversificationSurface}
\alias{plot.auteurMCMCMC}
\alias{plot.auteurMCMC}
\alias{plot.rjmcmc}
\alias{plot.rjmcmc.list}
\alias{print.auteurRAW}
\alias{print.constraint.m}
\alias{print.mcmc.list}
\alias{print.rbm}
\alias{print.rjmcmc}
\alias{print.rjmcmcmc}
\alias{print.taxdump}
\alias{print.transformer}
\alias{rcp}
\alias{resplitEdgeMatrixGeiger}
\alias{root.phylo}
\alias{sem}
\alias{set.seed.clock}
\alias{splitEdgeMatrixGeiger}
\alias{subset.phylo}
\alias{trace.plot}
\alias{transform}
\alias{ultrametricize.phylo}
\alias{unique.multiPhylo}
\alias{unique.phylo}
\alias{white.mkn}
\alias{write.pathd8}
\alias{write.r8s}
\alias{write.treePL}

\title{internal geiger functions}

\description{\code{geiger-internal} functions are either not typically called by the user or are currently undocumented}

\details{
This is an internal \pkg{geiger} function, either not intended to be called directly by the user or is currently undocumented.
}

