\name{geiger-internal}

\alias{print.bm}
\alias{print.mkn}
\alias{transform.phylo}
\alias{ROOT.ALL}
\alias{ROOT.BOTH}
\alias{ROOT.EQUI}
\alias{ROOT.FLAT}
\alias{ROOT.GIVEN}
\alias{ROOT.MAX}
\alias{ROOT.OBS}
\alias{acegram}
\alias{argnames}
\alias{argnames.bm}
\alias{argnames<-}
\alias{argnames<-.bm}
\alias{bf}
\alias{dlunif}
\alias{drop.tip}
\alias{dtpois}
\alias{edges.phylo}
\alias{edgelabels.auteur}
\alias{ess}
\alias{hash.node}
\alias{hashes.phylo}
\alias{hashes.rjmcmc}
\alias{hdr}
\alias{heights}
\alias{heights.multiPhylo}
\alias{heights.phylo}
\alias{hme}
\alias{is.phylo}
\alias{is.root}
\alias{likfx.bm}
\alias{make.bm.relaxed}
\alias{plot.auteurMULTI}
\alias{plot.auteurSINGLE}
\alias{plot.rjmcmc}
\alias{plot.rjmcmc.list}
\alias{print.auteurRAW}
\alias{print.mcmc.list}
\alias{print.rbm}
\alias{print.rjmcmc}
\alias{print.rjmcmcMULTI}
\alias{root.phylo}
\alias{sem}
\alias{PATHd8.phylo}
\alias{subset.phylo}
\alias{trace.plot}
\alias{ultrametricize.phylo}
\alias{unique.multiPhylo}
\alias{unique.phylo}
\alias{write.pathd8}
\alias{write.r8s}
\alias{write.treePL}

\title{internal geiger functions}

\description{\code{geiger-internal} functions are either not typically called by the user or are currently undocumented}

\details{
This is an internal \pkg{geiger} function, either not intended to be called directly by the user or is currently undocumented.
}

