\name{geiger-internal}

\alias{print.bm}
\alias{print.mkn}
\alias{argn}
\alias{argn.default}
\alias{argn.bm}
\alias{argn.mkn}
\alias{argn<-}
\alias{argn<-.bm}
\alias{as.phylo.hphylo}
\alias{bf}
\alias{bm.lik}
\alias{cherries}
\alias{coef.gfit}
\alias{coef.gfits}
\alias{constrain}
\alias{constrain.k}
\alias{constrain.m}
\alias{dlunif}
\alias{drop.tip}
\alias{dtpois}
\alias{edges.phylo}
\alias{edgelabels.auteur}
\alias{ess}
\alias{exemplar.phylo}
\alias{fitDiversification}
\alias{fitSplitModel}
\alias{gbresolve.default}
\alias{gbtaxdump}
\alias{gen}
\alias{getBD}
\alias{getDiversificationModel}
\alias{getFullSplitModel}
\alias{hash.node}
\alias{hashes.phylo}
\alias{hashes.rjmcmc}
\alias{hdr}
\alias{heights}
\alias{heights.multiPhylo}
\alias{heights.phylo}
\alias{hme}
\alias{is.constrained}
\alias{is.phylo}
\alias{is.root}
\alias{kendallmoran.rate}
\alias{likfx.bm}
\alias{Linnaean}
\alias{load}
\alias{logLik.gfit}
\alias{logLik.gfits}
\alias{make.bm.relaxed}
\alias{mkn.lik}
\alias{PATHd8.phylo}
\alias{plotDiversificationSurface}
\alias{plot.auteurMCMCMC}
\alias{plot.auteurMCMC}
\alias{plot.rjmcmc}
\alias{plot.rjmcmc.list}
\alias{print.auteurRAW}
\alias{print.bayesfactor}
\alias{print.constraint.m}
\alias{print.glnL}
\alias{print.gprior}
\alias{print.mcmc.list}
\alias{print.rbm}
\alias{print.rjmcmc}
\alias{print.rjmcmcmc}
\alias{print.taxdump}
\alias{print.transformer}
\alias{resplitEdgeMatrixGeiger}
\alias{root.phylo}
\alias{sim.mecca}
\alias{sem}
\alias{span.phylo}
\alias{splitEdgeMatrixGeiger}
\alias{ultrametricize.phylo}
\alias{unique.multiPhylo}
\alias{unique.phylo}
\alias{white.mkn}
\alias{write.pathd8}
\alias{write.r8s}
\alias{write.treePL}

\title{internal geiger functions}

\description{\code{geiger-internal} functions are either not typically called by the user or are currently undocumented}

\details{
This is an internal \pkg{geiger} function, either not intended to be called directly by the user or is currently undocumented.
}

