\name{geiger-package}
\alias{geiger-package}
\alias{geiger}
\docType{package}
\title{
GEIGER
}
\description{
 A package for macroevolutionary simulation and estimating parameters 
 related to diversification from comparative phylogenetic data.
}
\details{
\tabular{ll}{
Package: \tab geiger\cr
Type: \tab Package\cr
License: \tab GPL version 2 or greater?\cr
}

}
\author{
LJ Harmon, J Weir, C Brock, RE Glor, W Challenger, G Hunt, R FitzJohn, MW Pennell, GJ Slater, JW Brown, J Uyeda, and JM Eastman

Maintainer: Matt Pennell <mwpennell@gmail.com>
}
