\name{geiger-package}
\alias{geiger-package}
\alias{geiger}
\docType{package}
\title{
GEIGER
}
\description{
 A package for macroevolutionary simulation and estimating parameters 
 related to diversification from comparative phylogenetic data.
}
\details{
\tabular{ll}{
Package: \tab geiger\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-01-19\cr
License: \tab GPL version 2 or greater?\cr
}

}
\author{
LJ Harmon, J Weir, C Brock, RE Glor, W Challenger, G Hunt, R FitzJohn, MW Pennell, GJ Slater, JW Brown, and JM Eastman

Maintainer: Luke Harmon <lukeh@uidaho.edu>
}
